Subscriber refactor
