+++ echb/newsevents/models/__init__.py
@@ -1,3 +1,5 @@
 from .author import Author
 from .event import Event
 from .news_item import NewsItem
+from .subscriber import Subscriber
+from .mailinglog import MailingLog
