+++ echb/pages/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 
 from django_summernote.admin import SummernoteModelAdmin
 
-from .models import Page, Feedback, Subscriber, OldUser, MailingLog
+from .models import Page, Feedback, OldUser
 
 
 @admin.register(Page)
@@ -16,10 +16,4 @@ class FeedbackAdmin(admin.ModelAdmin):
     list_display = ['subject', 'email', 'name', 'created']
 
 
-@admin.register(Subscriber)
-class SubscriberAdmin(admin.ModelAdmin):
-    list_display = ['email', 'activated', 'created']
-
-
 admin.site.register(OldUser)
-admin.site.register(MailingLog)
