+++ echb/pages/forms.py
@@ -1,10 +1,8 @@
 from django.forms import ModelForm
 from django.core import mail
 from django import forms
-from django.contrib.sites.shortcuts import get_current_site
-from django.template.loader import get_template
 
-from .models import Feedback, Subscriber
+from .models import Feedback
 
 
 class FeedbackForm(ModelForm):
@@ -49,27 +47,3 @@ class FeedbackForm(ModelForm):
         labels = {
             'cc_myself': 'Отправить копию сообщения мне'
         }
-
-
-class SubscriberForm(ModelForm):
-    def get_domain(self, request):
-        current_site = get_current_site(request)
-        return current_site.domain
-
-    def send_mail(self, subscriber, domain):
-        email = self.cleaned_data['email']
-        message = get_template(
-            'pages/subscriber_activation_letter.html').render({'subscriber': subscriber, 'domain': domain})
-        with mail.get_connection() as connection:
-            email = mail.EmailMessage(
-                subject='Подтверждение о подписке на новости',
-                body=message,
-                to=(email,),
-                connection=connection
-            )
-            email.content_subtype = 'html'
-            email.send()
-
-    class Meta:
-        model = Subscriber
-        fields = ['email']
