+++ echb/pages/models.py
@@ -1,7 +1,6 @@
 from django.utils import timezone
 from django.db import models
 from django.urls import reverse
-import uuid
 from helpers.models import Audit, Seo
 
 
@@ -40,19 +39,6 @@ class Feedback(Audit):
         ordering = ['created']
 
 
-class Subscriber(Audit):
-    email = models.EmailField(unique=True)
-    activated = models.BooleanField(default=False)
-    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-
-    def __str__(self):
-        return self.email
-
-    class Meta:
-        verbose_name = 'подписчик'
-        verbose_name_plural = 'подписчики'
-
-
 class OldUser(models.Model):
     login = models.CharField(max_length=100)
     first_name = models.CharField(max_length=150)
@@ -61,18 +47,3 @@ class OldUser(models.Model):
 
     def __str__(self):
         return self.email
-
-
-class MailingLog(models.Model):
-    date = models.DateField(default=timezone.now)
-    emails = models.TextField(null=True, blank=True)
-    message = models.TextField(null=True, blank=True)
-    started = models.TimeField(default=timezone.now)
-    finished = models.TimeField(default=timezone.now)
-
-    def __str__(self):
-        return f'Рассылка от {self.date.strftime("%d/%m/%y")}'
-
-    class Meta:
-        verbose_name = 'лог рассылки'
-        verbose_name_plural = 'логи рассылок'
