+++ echb/pages/tests.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 from django.utils import timezone
 from django.core import mail
 
-from ..models import Page, Subscriber
+from .models import Page
 from newsevents.models import NewsItem, Event, Author
 from articles.models import Article, Category, Author as ArticleAuthor
 from galleries.models import Gallery, Author as GalleryAuthor
@@ -68,74 +68,6 @@ class PagesTests(TestCase):
     def test_home_page_contains_subscriber_form(self):
         response = self.client.get(reverse('home-page-unique'))
         self.assertContains(response, 'form')
-
-    def test_user_can_subscribe_to_letter(self):
-        response = self.client.post(
-            reverse('home-page-unique'), data={'email': 'test@test.ru'})
-        self.assertEqual(len(mail.outbox), 1)
-        self.assertContains(
-            response, 'Ваш email добавлен в список подписчиков')
-
-    def test_subscriber_not_added_if_email_incorrect(self):
-        self.client.post(
-            reverse('home-page-unique'), data={'email': 'test'})
-        self.assertEqual(len(mail.outbox), 0)
-
-    def test_subscriber_can_activate_his_subscription(self):
-        self.client.post(
-            reverse('home-page-unique'), data={'email': 'test@test.ru'})
-        self.assertEqual(mail.outbox[0].subject, 'Подтверждение о подписке на новости')
-        subscriber = Subscriber.objects.all().first()
-        self.assertFalse(subscriber.activated)
-        change_url = reverse('activate-subscriber', kwargs={'uuid': subscriber.uuid})
-        self.assertIn(change_url, mail.outbox[0].body)
-        self.client.get(reverse('activate-subscriber', kwargs={'uuid': subscriber.uuid}))
-        subscriber = Subscriber.objects.all().first()
-        self.assertTrue(subscriber.activated)
-
-    def test_subscriber_not_activated_with_wrong_uuid(self):
-        self.client.post(
-            reverse('home-page-unique'), data={'email': 'test@test.ru'})
-        subscriber = Subscriber.objects.all().first()
-        self.assertFalse(subscriber.activated)
-        subscriber_uuid = '98194856-4050-4397-9388-396669b5485b'
-        self.client.get(reverse('activate-subscriber', kwargs={'uuid': subscriber_uuid}))
-        subscriber = Subscriber.objects.all().first()
-        self.assertFalse(subscriber.activated)
-
-    def test_after_subscription_user_get_last_letter(self):
-        self.client.post(
-            reverse('home-page-unique'), data={'email': 'test@test.ru'})
-        self.assertEqual(mail.outbox[0].subject, 'Подтверждение о подписке на новости')
-        subscriber = Subscriber.objects.all().first()
-        self.client.get(reverse('activate-subscriber', kwargs={'uuid': subscriber.uuid}))
-
-        subscriber = Subscriber.objects.all().first()
-        self.assertTrue(subscriber.activated)
-
-        self.assertEqual(mail.outbox[1].subject, 'Последние новости с сайта ecb.kh.ua')
-        self.assertEqual(len(mail.outbox), 2)
-
-    def test_send_several_letters(self):
-        for item in range(10):
-            subscriber = Subscriber.objects.create(email=f'test_{item}@test.ru', activated=True)
-            subscriber.save()
-
-        self.client.get(reverse('send_letter'))
-        self.assertEqual(len(mail.outbox), 10)
-
-    def test_letter_was_already_sent_recently(self):
-        for item in range(10):
-            subscriber = Subscriber.objects.create(email=f'test_{item}@test.ru', activated=True)
-            subscriber.save()
-
-        response = self.client.get(reverse('send_letter'))
-        self.assertContains(response, 'Letters were successfuly sent')
-
-        response = self.client.get(reverse('send_letter'))
-        self.assertContains(response, 'Letters were sent earlier')
-
-        self.assertEqual(len(mail.outbox), 10)
 
     def test_about_us_page_contains_menu(self):
         Page.objects.create(
