+++ echb/pages/views.py
@@ -1,28 +1,19 @@
 import logging
 import os
-from datetime import datetime, timedelta
-
 import requests
-from django.contrib import messages
-from django.contrib.auth.decorators import login_required
-from django.contrib.sites.shortcuts import get_current_site
-from django.core import mail
-from django.http import HttpResponse, HttpResponseRedirect
-from django.contrib.auth.models import User
-from django.shortcuts import redirect, render, render_to_response, reverse
-from django.template.loader import get_template
-from django.utils import timezone
-from django.views.generic import DetailView, FormView, ListView
-from django.views.generic.edit import ModelFormMixin
-from django.views.generic.base import TemplateView, View
+from datetime import datetime
+
+from django.shortcuts import redirect, render, render_to_response
+from django.views.generic import DetailView, FormView
+from django.views.generic.base import View, TemplateView
 
 from articles.models import Article
 from galleries.models import Gallery
 from newsevents.models import Event, NewsItem
+from newsevents.forms import SubscriberForm
 
-from .forms import FeedbackForm, SubscriberForm
-from .models import MailingLog, Page, Subscriber
-
+from .forms import FeedbackForm
+from .models import Page
 logger = logging.getLogger('ECHB')
 
 
@@ -96,108 +87,6 @@ class ContactsThankYouView(TemplateView):
     template_name = 'pages/thankyou.html'
 
 
-class ActivateSubscriber(View):
-    def get(self, request, uuid):
-        subscriber = Subscriber.objects.filter(uuid=uuid).first()
-
-        if subscriber:
-            subscriber.activated = True
-            subscriber.save()
-
-            send_mail(request, [subscriber.email, ])
-
-            return redirect('subscriber-activated')
-        return redirect('home')
-
-
-class SendLetterToSubscribers(View):
-    def get(self, request):
-        last_month = datetime.now() - timedelta(days=30)
-        already_sent = MailingLog.objects.filter(date__gte=last_month).filter(date__lte=datetime.now()).first()
-
-        if already_sent is not None:
-            return HttpResponse(f'Letters were sent earlier: {already_sent.date.strftime("%d/%m/%y")}',
-                                content_type="text/plain")
-
-        subscribers = Subscriber.objects.filter(activated=True)
-
-        emails = [subscriber.email for subscriber in subscribers]
-        send_mail(request, emails)
-        return HttpResponse(f'Letters were successfuly sent: {timezone.now().strftime("%d/%m/%y")}',
-                            content_type="text/plain")
-
-
-def get_context_for_letter(request):
-    time_delta_past = datetime.today() - timedelta(days=30)
-    time_delta_future = datetime.today() + timedelta(days=30)
-    news = NewsItem.objects.filter(publication_date__gte=time_delta_past).filter(
-        publication_date__lt=datetime.today()).filter(published=True)
-    events = Event.objects.filter(date__gt=datetime.today()).filter(date__lt=time_delta_future)
-    articles = Article.objects.filter(date__gte=time_delta_past).filter(
-        date__lt=datetime.today()).select_related('category').select_related('author')
-    domain = get_domain(request)
-    context = {
-        'news': news,
-        'events': events,
-        'articles': articles,
-        'domain': domain
-    }
-    return context
-
-
-class Letter(View):
-    def get(self, request):
-        context = get_context_for_letter(request)
-        return render(request, 'newsevents/letter.html', context)
-
-
-def send_mail(request, emails):
-    mailing_log = MailingLog()
-    mailing_log.save()
-
-    context = get_context_for_letter(request)
-    message = get_template('newsevents/letter.html').render(context)
-
-    emails_for_log = emails[:]  # to change array if there is any error with email
-
-    for address in emails:
-        try:
-            with mail.get_connection() as connection:
-                email = mail.EmailMessage(
-                    subject='Последние новости с сайта ecb.kh.ua',
-                    body=message,
-                    to=(address,),
-                    connection=connection
-                )
-                email.content_subtype = 'html'
-                email.send()
-        except Exception as e:
-            emails_for_log.pop(address, None)
-            logger.error(f'Errors on newsletter sending: {e.args}')
-
-    mailing_log_finished = MailingLog.objects.filter(pk=mailing_log.id).first()
-    mailing_log_finished.message = message
-    mailing_log_finished.finished = timezone.now()
-    mailing_log_finished.emails = emails_for_log
-    mailing_log_finished.save()
-
-    return mailing_log_finished
-
-
-def get_domain(request):
-    current_site = get_current_site(request)
-    return f'{request.scheme}://{current_site.domain}'
-
-
-def check_captcha(request):
-    captcha = request.POST.get('g-recaptcha-response')
-    response = requests.post("https://www.google.com/recaptcha/api/siteverify",
-                             data={'secret': '6LfamGAUAAAAAEnS0-AF5p_EVmAFriMZqkkll-HM', 'response': captcha})
-
-    debug_mode = bool(os.environ.get("DEBUG", False))
-    return debug_mode if debug_mode else response.json()['success']
-
-
 def handler404(request, exception, template_name='pages/404.html'):
     response = render_to_response('pages/404.html')
     response.status_code = 404
@@ -208,3 +97,12 @@ def handler500(request):
     response = render_to_response('pages/500.html')
     response.status_code = 500
     return response
+
+
+def check_captcha(request):
+    captcha = request.POST.get('g-recaptcha-response')
+    response = requests.post("https://www.google.com/recaptcha/api/siteverify",
+                             data={'secret': '6LfamGAUAAAAAEnS0-AF5p_EVmAFriMZqkkll-HM', 'response': captcha})
+
+    debug_mode = bool(os.environ.get("DEBUG", False))
+    return debug_mode if debug_mode else response.json()['success']
